code/requirements.txt
# [INFO] conditional installations: https://peps.python.org/pep-0496/

# General
setuptools
pip
wheel

# numpy
numpy<1.22,>=1.18
sympy<1.10

# cuda nvidia
# --extra-index-url https://pypi.ngc.nvidia.com
# nvidia-pyindex; sys_platform != 'darwin'
# nvidia-cuda-runtime-cu11; sys_platform != 'darwin'

# Install/Upgrade Pennylane
pennylane

# Install Pennylane Plugins
pennylane-sf
pennylane-qiskit
amazon-braket-pennylane-plugin
pennylane-cirq
pennylane-forest
pennylane-qsharp

# Install Interfaces
autograd
tensorflow>=1.13.2
jax
jaxlib
torch
torchvision
torchaudio

# Install Amazon Braket Python SDK
amazon-braket-sdk

# Install/Upgrade boto3
boto3

# Install/Upgrade sklearn
sklearn

# Installing matplotlip draw library
pylatexenc

# Installing qiskit machine learning library
qiskit-machine-learning
qiskit-optimization
qiskit-aer-gpu
# qiskit-aer; sys_platform == 'darwin'

# Installing utilities
tqdm
GPUtil

# Installing pydot
pydot

# Adding additional libraries for notebook tooling
git+https://github.com/qiskit-community/qiskit-textbook.git#subdirectory=qiskit-textbook-src
numexpr

